feat: 면접 종료 후 결과 보내는 기능 위치 변경
